re-used _resolve_tag_to_tei_mapping_namespace
